requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.0
scipy==1.11.4
spacy==3.7.2
python-docx==0.8.11
PyMuPDF==1.23.5